fixed missing config error in setup scripts

+++ privoxy-shared-setup.sh
@@ -4,7 +4,9 @@
 
 curl -o "/usr/local/etc/privoxy/ppilot.sh" "https://raw.githubusercontent.com/brian-beeler/privoxy-pilot-macos/main/ppilot.sh"
 chmod og+x /usr/local/etc/privoxy/ppilot.sh
-mv /usr/local/etc/privoxy/config /usr/local/etc/privoxy/config.original
+if [ -f "/usr/local/etc/privoxy/config" ]; then
+  mv /usr/local/etc/privoxy/config /usr/local/etc/privoxy/config.original
+fi
 echo -e "\r\n\r\n# \r\n# \r\n# \r\n# " >> /usr/local/etc/privoxy/config
 echo -e "# allow privoxy to make connections with the local network" >> /usr/local/etc/privoxy/config
 echo -e "listen-address $(ifconfig | grep "inet " | grep -v 127.0.0.1 | awk '{print $2}')":8118 >> /usr/local/etc/privoxy/config
